Revamped thread sleeping to be handled by the scheduler.
Prior to this change, a thread would context switch back into the thread sleep function every time it got scheduled to run, even though it is fairly obvious that the thread is sleeping. Now thread sleeps are handled by a suspension tick counter in the scheduler so the thread will not wake up until the timer elapses.