Tracked down the final long running hardfault.
Interrupts from peripherals with a higher priority than pendSV will cause issues as upon the re-enabling of interrupts in pendSV_Handler,
it may interrupt to a different handler for a peripheral. This can bork the context switch as the final BX instruction is missed.

TODO: look into priority grouping to avoid nested interrupts from causing issues!